+++ src/Interop/LibHeif/NativeMethods/LibHeifNative.Encoder.cs
@@ -46,6 +46,15 @@ namespace LibHeifSharp.Interop
                                                                                                 ref int maximum);
 
         [DllImport(DllName, CallingConvention = DllCallingConvention)]
+        internal static extern unsafe heif_error heif_encoder_parameter_get_valid_integer_values(heif_encoder_parameter parameter,
+                                                                                                 [MarshalAs(UnmanagedType.Bool)] out bool haveMinimum,
+                                                                                                 [MarshalAs(UnmanagedType.Bool)] out bool haveMaximum,
+                                                                                                 ref int minimum,
+                                                                                                 ref int maximum,
+                                                                                                 out int numValidValues,
+                                                                                                 out IntPtr validValuesArray);
+
+        [DllImport(DllName, CallingConvention = DllCallingConvention)]
         internal static extern unsafe heif_error heif_encoder_parameter_get_valid_string_values(heif_encoder_parameter parameter,
                                                                                                 out OutputStringArray array);